execution/example_script.py
#!/usr/bin/env python3
"""
Example Execution Script

This is a template for execution layer scripts.
Purpose: [Describe what this script does]
Input: [Describe input parameters]
Output: [Describe output format and location]
"""

import os
import sys
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

def main():
    """
    Main execution function.
    This is where the deterministic logic goes.
    """
    try:
        # Example: Get API key from environment
        api_key = os.getenv('API_KEY')
        if not api_key:
            raise ValueError("API_KEY not found in environment variables")
        
        # Your deterministic logic here
        print("Script executed successfully")
        
        return True
        
    except Exception as e:
        print(f"Error: {str(e)}", file=sys.stderr)
        sys.exit(1)

if __name__ == "__main__":
    main()
