requirements.txt
streamlit>=1.30.0
langgraph>=0.0.30
langchain-core>=0.1.0
langchain-openai>=0.0.5
langchain-groq>=0.1.0
plotly>=5.18.0
fastapi>=0.108.0
uvicorn>=0.25.0
requests>=2.31.0
python-dotenv>=1.0.0
pandas>=2.0.0
openai>=1.0.0
groq>=0.4.0
